Rename Java package for API Tables ITs (to conform to Java package naming convention)
